+++ Task043.cs
@@ -6,9 +6,9 @@ namespace euler_from26
     public static class Task43
     {
 
-        public static BigInteger d(int[] ints, params int[] i)
+        public static BigInteger d(long[] ints, params long[] i)
         {
-            int[] z = new int[i.Length];
+            long[] z = new long[i.Length];
             for (int j = 0; j < i.Length; j++)
                 z[i.Length - j - 1] = ints[i[j]];
             return Digits.int_from_digits(Functions.ToList(z));
@@ -19,15 +19,15 @@ namespace euler_from26
             // foreach (var p in Functions.Permute(new int[]{0, 1, 2}, 0))
             //     Console.WriteLine($"{MyCollections.Print(p)}\t{Digits.int_from_digits(p, true)}");
             BigInteger sum = 0;
-            foreach(var p in Functions.Permute(new int[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9}, 0))
+            foreach(var p in Functions.Permute(new long[]{0, 1, 2, 3, 4, 5, 6, 7, 8, 9}, 0))
             {
-                var d234 = d(p, new int[]{1, 2, 3});
-                var d345 = d(p, new int[]{2, 3, 4});
-                var d456 = d(p, new int[]{3, 4, 5});
-                var d567 = d(p, new int[]{4, 5, 6});
-                var d678 = d(p, new int[]{5, 6, 7});
-                var d789 = d(p, new int[]{6, 7, 8});
-                var d8910 = d(p, new int[]{7, 8, 9});
+                var d234 = d(p, new long[]{1, 2, 3});
+                var d345 = d(p, new long[]{2, 3, 4});
+                var d456 = d(p, new long[]{3, 4, 5});
+                var d567 = d(p, new long[]{4, 5, 6});
+                var d678 = d(p, new long[]{5, 6, 7});
+                var d789 = d(p, new long[]{6, 7, 8});
+                var d8910 = d(p, new long[]{7, 8, 9});
                 if (    Prime.isDivisible(d234, 2)
                     && Prime.isDivisible(d345, 3)
                     && Prime.isDivisible(d456, 5)
