+++ Task079.cs
@@ -6,7 +6,7 @@ namespace euler_from26
     public static class Task079
     {
 
-        public static bool suits(int[] perm, int[]pass)
+        public static bool suits(long[] perm, long[]pass)
         {
             int i1 = Array.IndexOf(perm, pass[0]);
             int i2 = Array.IndexOf(perm, pass[1]);
@@ -14,7 +14,7 @@ namespace euler_from26
             return i1 < i2 && i2 < i3;
         }
 
-        public static bool suits(int[] perm, List<int[]>all)
+        public static bool suits(long[] perm, List<long[]>all)
         {
             foreach(var p in all)
                 if (!suits(perm, p))
@@ -33,10 +33,10 @@ namespace euler_from26
                 foreach(var d in Digits.digits(p))
                     digits.Add(d);
             }
-            List<int[]> distinct_keys_array = new();
+            List<long[]> distinct_keys_array = new();
             foreach(var ddd in pass)
                 distinct_keys_array.Add(Digits.digitsAsArray(ddd));
-            int[] digits_array = new int[digits.Count];
+            long[] digits_array = new long[digits.Count];
             int i = 0;
             foreach(int d in digits)
                 digits_array[i++] = d;
